+++ cypress-tests/cypress/e2e/configs/Payment/Trustpay.js
@@ -64,7 +64,7 @@ export const connectorDetails = {
         body: {
           status: "requires_payment_method",
           shipping_cost: 50,
-          amount: 6500,
+          amount: 6000,
         },
       },
     },
@@ -82,9 +82,9 @@ export const connectorDetails = {
         body: {
           status: "succeeded",
           shipping_cost: 50,
-          amount_received: 6550,
-          amount: 6500,
-          net_amount: 6550,
+          amount_received: 6050,
+          amount: 6000,
+          net_amount: 6050,
         },
       },
     },
@@ -142,12 +142,7 @@ export const connectorDetails = {
         },
       },
       Request: {
-        payment_method: "card",
-        payment_method_data: {
-          card: successfulNo3DSCardDetails,
-        },
-        currency: "USD",
-        customer_acceptance: null,
+        amount_to_capture: 6000,
       },
       Response: {
         status: 400,
@@ -173,16 +168,7 @@ export const connectorDetails = {
         },
       },
       Request: {
-        payment_method: "card",
-        payment_method_data: {
-          card: successfulNo3DSCardDetails,
-        },
-        currency: "USD",
-        paymentSuccessfulStatus: "succeeded",
-        paymentSyncStatus: "succeeded",
-        refundStatus: "succeeded",
-        refundSyncStatus: "succeeded",
-        customer_acceptance: null,
+        amount_to_capture: 2000,
       },
       Response: {
         status: 400,
@@ -213,12 +199,7 @@ export const connectorDetails = {
         },
       },
       Request: {
-        payment_method: "card",
-        payment_method_data: {
-          card: successfulNo3DSCardDetails,
-        },
-        currency: "USD",
-        customer_acceptance: null,
+        amount: 6000,
       },
       Response: {
         status: 200,
@@ -235,12 +216,7 @@ export const connectorDetails = {
         },
       },
       Request: {
-        payment_method: "card",
-        payment_method_data: {
-          card: successfulNo3DSCardDetails,
-        },
-        currency: "USD",
-        customer_acceptance: null,
+        amount: 2000,
       },
       Response: {
         status: 200,
@@ -256,14 +232,6 @@ export const connectorDetails = {
           STATUS: true,
           TIMEOUT: 15000,
         },
-      },
-      Request: {
-        payment_method: "card",
-        payment_method_data: {
-          card: successfulNo3DSCardDetails,
-        },
-        currency: "USD",
-        customer_acceptance: null,
       },
       Response: {
         status: 200,